+++ src/core/QueryManager.ts
@@ -500,7 +500,16 @@ export class QueryManager {
     return resPromise;
   }
 
-  public fetchQuery<T>(queryId: string, options: WatchQueryOptions, fetchType?: FetchType): Promise<ApolloQueryResult<T>> {
+  public fetchQuery<T>(
+    queryId: string,
+    options: WatchQueryOptions,
+    fetchType?: FetchType,
+
+    // This allows us to track if this is a query spawned by a `fetchMore`
+    // call for another query. We need this data to compute the `fetchMore`
+    // network status for the query this is fetching for.
+    fetchMoreForQueryId?: string,
+  ): Promise<ApolloQueryResult<T>> {
     const {
       variables = {},
       forceFetch = false,
@@ -554,6 +563,7 @@ export class QueryManager {
       storePreviousVariables: shouldFetch,
       isPoll: fetchType === FetchType.poll,
       isRefetch: fetchType === FetchType.refetch,
+      fetchMoreForQueryId,
       metadata,
     });
 
@@ -577,6 +587,7 @@ export class QueryManager {
         queryId,
         document: queryDoc,
         options,
+        fetchMoreForQueryId,
       });
     }
 
@@ -918,11 +929,13 @@ export class QueryManager {
     queryId,
     document,
     options,
+    fetchMoreForQueryId,
   }: {
     requestId: number,
     queryId: string,
     document: DocumentNode,
     options: WatchQueryOptions,
+    fetchMoreForQueryId?: string,
   }): Promise<ExecutionResult> {
     const {
       variables,
@@ -951,6 +964,7 @@ export class QueryManager {
             result,
             queryId,
             requestId,
+            fetchMoreForQueryId,
             extraReducers,
           });
 
@@ -1003,6 +1017,7 @@ export class QueryManager {
               error,
               queryId,
               requestId,
+              fetchMoreForQueryId,
             });
 
             this.removeFetchQueryPromise(requestId);